Add options to compute stats via command line
Added three arguments invoked via command line to control the
diags to be computed:

   -diff_rms will compute horizontal mean difference and rms
    hovmuller plots (time vs depth)
   -surface will compute stats for surface variables (these
    variables are hard-coded for now)
   -forcing will compute stats for forcing variables (also
    hard-coded for now)

In all options, a dataArray with region masks (regions are hard-coded
for now) is provided, so the diagnostics are computed for each region.

The calculations are done using Dask and the chuncking and number
of works are hard-coded.

TODO:
* Add option to specify variables to be processed (either via cmd line
of yaml file);
* Optimize the selection of workers and chuncking automatically.