Add non_local option
There is a new argument (no_local) in the latest verion of CIME
that needs to be set to true when when you've requested a machine
that you aren't on. This PR adds this new argument. I am not sure
if this will break mom6-tools for older versions of CIME.