Add new API to compute drift and drift RMS
Script drift.py replaces diff_rms.py.
The new version computes horizontal mean difference (drift) and RMS
for a generic tracer instead for a hard-coded set of tracers (as in
the old version). This new approach speeds up the generation of
diagnostics because the same script can be invoked in parallel
for multiple tracers.